+++ benchmark/benchmark_utils.py
@@ -8,7 +8,7 @@ from typing import Dict, Any
 import torch
 import onnxruntime as ort
 
-from benchmark import PyTorchBenchmark, ONNXBenchmark, OVBenchmark
+from benchmark_class import PyTorchBenchmark, ONNXBenchmark, OVBenchmark
 
 
 def run_benchmark(
